refactor(mobile): extract terminal touch controller
