+++ packages/app/src/components/terminal.tsx
@@ -17,9 +17,24 @@ import { useTerminal, type LocalPTY } from "@/context/terminal"
 import { terminalAttr, terminalProbe } from "@/testing/terminal"
 import { disposeIfDisposable, getHoveredLinkText, setOptionIfSupported } from "@/utils/runtime-adapters"
 import { terminalWriter } from "@/utils/terminal-writer"
+import { useTerminalUiBindings } from "@/components/terminal-touch-controller"
 
 const TOGGLE_TERMINAL_ID = "terminal.toggle"
 const DEFAULT_TOGGLE_TERMINAL_KEYBIND = "ctrl+`"
+
+export interface TerminalSelectionApi {
+  // Deliberately NOT delegating to `term.hasSelection()`: that method hides
+  // an in-progress selection until its internal drag threshold is met,
+  // which would report `false` right after a long-press-only word-select
+  // (see beginSelection in useTerminalUiBindings) even though text is
+  // already highlighted and copied. Reading the text directly sidesteps
+  // that gate and matches what's actually visible on screen.
+  hasSelection(): boolean
+  copySelection(): boolean
+  paste(text: string): void
+  onSelectionChange(cb: () => void): VoidFunction
+}
+
 export interface TerminalProps extends ComponentProps<"div"> {
   pty: LocalPTY
   autoFocus?: boolean
@@ -28,6 +43,7 @@ export interface TerminalProps extends ComponentProps<"div"> {
   onConnect?: () => void
   onConnectError?: (error: unknown) => void
   onSend?: (fn: ((data: string) => void) | undefined) => void
+  onSelectionApi?: (api: TerminalSelectionApi | undefined) => void
 }
 
 let sharedModule: Promise<typeof import("ghostty-web")> | undefined
@@ -205,310 +221,6 @@ const debugTerminal = (...values: unknown[]) => {
   console.debug("[terminal]", ...values)
 }
 
-const errorName = (err: unknown) => {
-  if (!err || typeof err !== "object") return
-  if (!("name" in err)) return
-  const errorName = err.name
-  return typeof errorName === "string" ? errorName : undefined
-}
-
-const useTerminalUiBindings = (input: {
-  container: HTMLDivElement
-  term: Term
-  cleanups: VoidFunction[]
-  handlePointerDown: (e: PointerEvent) => void
-  handleLinkClick: (event: MouseEvent) => void
-}) => {
-  const handleCopy = (event: ClipboardEvent) => {
-    const selection = input.term.getSelection()
-    if (!selection) return
-
-    const clipboard = event.clipboardData
-    if (!clipboard) return
-
-    event.preventDefault()
-    clipboard.setData("text/plain", selection)
-  }
-
-  const handlePaste = (event: ClipboardEvent) => {
-    const clipboard = event.clipboardData
-    const text = clipboard?.getData("text/plain") ?? clipboard?.getData("text") ?? ""
-    if (!text) return
-
-    event.preventDefault()
-    event.stopPropagation()
-    input.term.paste(text)
-  }
-
-  const handleTextareaFocus = () => {
-    input.term.options.cursorBlink = true
-  }
-  const handleTextareaBlur = () => {
-    input.term.options.cursorBlink = false
-  }
-
-  input.container.addEventListener("copy", handleCopy, true)
-  input.cleanups.push(() => input.container.removeEventListener("copy", handleCopy, true))
-
-  input.container.addEventListener("paste", handlePaste, true)
-  input.cleanups.push(() => input.container.removeEventListener("paste", handlePaste, true))
-
-  input.container.addEventListener("pointerdown", input.handlePointerDown)
-  input.cleanups.push(() => input.container.removeEventListener("pointerdown", input.handlePointerDown))
-
-  // --- mobile touch: drag-to-scroll (pointermove swipe detection) ---
-  // We do NOT intercept touchend or pointerup — Ghostty's internal
-  // `canvas.addEventListener("touchend", …)` handler is allowed to fire
-  // (it calls `textarea.focus()` which is the ONLY reliable way to attach
-  // the Android softkeyboard IME to the hidden textarea). Tap therefore
-  // behaves exactly as on HEAD (opens keyboard), and we only add a new
-  // pointermove-driven swipe gesture that scrolls the scrollback in place.
-  const MOBILE_SWIPE_THRESHOLD_PX = 8
-
-  // Ghostty's own `scrollLines` clamps at the buffer's live bottom (viewportY
-  // 0) — dragging further has no effect, so the last prompt is stuck at the
-  // bottom edge of the panel even when the on-screen keyboard covers most of
-  // it. Fake "scrolling past the end" by translating the (already fully
-  // rendered) canvas upward via CSS once the real scrollback is exhausted —
-  // `getCanvasOffset()`/hit-testing inside ghostty-web reads
-  // `canvas.getBoundingClientRect()`, which already reflects a CSS
-  // transform, so tap/selection coordinates stay correct while overscrolled.
-  const MAX_OVERSCROLL_FACTOR = 0.5
-  let overscrollPx = 0
-  const canvasEl = input.container.querySelector("canvas")
-
-  const applyOverscroll = () => {
-    if (!canvasEl) return
-    canvasEl.style.transform = overscrollPx > 0 ? `translateY(-${overscrollPx}px)` : ""
-  }
-
-  // Overscroll is deliberately persistent — it must survive both incoming
-  // shell output and the user's own typing, since the whole point is keeping
-  // the last prompt visible above the on-screen keyboard while working. It
-  // only changes via the drag-to-unwind path in onTouchMoveCapture below.
-  //
-  // The keyboard opening/closing still resizes `container` (via the
-  // --vv-top/--vvh compensation), which re-fits the terminal grid. Since
-  // `overscrollPx` is stored as an absolute pixel amount computed against
-  // `container`'s height at drag time, if the keyboard then shrinks that
-  // container, the same absolute offset becomes disproportionately large
-  // relative to the new (smaller) height. Re-clamp it live so it never
-  // exceeds MAX_OVERSCROLL_FACTOR of whatever the container's current height
-  // is — this is the only thing that adjusts the offset outside a drag.
-  const clampOverscrollToContainer = () => {
-    const maxOverscrollPx = input.container.clientHeight * MAX_OVERSCROLL_FACTOR
-    if (overscrollPx <= maxOverscrollPx) return
-    overscrollPx = Math.max(0, maxOverscrollPx)
-    applyOverscroll()
-  }
-  if (typeof ResizeObserver !== "undefined") {
-    const overscrollResizeObserver = new ResizeObserver(clampOverscrollToContainer)
-    overscrollResizeObserver.observe(input.container)
-    input.cleanups.push(() => overscrollResizeObserver.disconnect())
-  }
-
-  type TouchMode = "pending" | "swipe"
-  let currentTouch: { id: number; x: number; y: number; mode: TouchMode; scrollApplied: number } | null = null
-  // Chromium fires `pointerup` before `touchend` for the same gesture
-  // (confirmed on-device: ~0.2ms apart, consistently pointerup-first across
-  // every touch sequence captured). `onTouchEndOrCancel` used to null
-  // `currentTouch` on pointerup, so by the time `blockTouchEndIfSwipe` ran
-  // on the later touchend, `currentTouch?.mode` always read as `undefined`
-  // — the swipe check never matched and scrolling always opened the
-  // keyboard. Capture the verdict before nulling so touchend can still see it.
-  let lastGestureWasSwipe = false
-
-  // Safety net: if a real pointerup/pointercancel is ever missed for the
-  // tracked pointer (observed as a risk, not directly reproduced: Android can
-  // drop a touch sequence without delivering either event to the WebView
-  // when the view hierarchy resizes mid-touch, e.g. the keyboard showing or
-  // hiding during a drag), `currentTouch` would otherwise stay non-null
-  // forever. Combined with `blockTouchEndIfSwipe`'s `|| currentTouch` guard
-  // below, a stuck `currentTouch` permanently stops Ghostty's own touchend
-  // handler from ever firing again for this terminal instance — the tab
-  // would look like it silently stopped accepting taps/keyboard focus, with
-  // no way to recover short of creating a new tab. Auto-clear it if no
-  // pointer activity refreshes this watchdog for a second; a real gesture
-  // always finishes (or keeps moving) well within that window.
-  const STUCK_TOUCH_TIMEOUT_MS = 1000
-  let stuckTouchTimer: ReturnType<typeof setTimeout> | undefined
-  const armStuckTouchWatchdog = () => {
-    if (stuckTouchTimer !== undefined) clearTimeout(stuckTouchTimer)
-    stuckTouchTimer = setTimeout(() => {
-      stuckTouchTimer = undefined
-      if (!currentTouch) return
-      lastGestureWasSwipe = currentTouch.mode === "swipe"
-      currentTouch = null
-    }, STUCK_TOUCH_TIMEOUT_MS)
-  }
-  const disarmStuckTouchWatchdog = () => {
-    if (stuckTouchTimer === undefined) return
-    clearTimeout(stuckTouchTimer)
-    stuckTouchTimer = undefined
-  }
-  input.cleanups.push(disarmStuckTouchWatchdog)
-
-  const mobileCharHeight = () => {
-    const rows = input.term.rows || 24
-    return Math.max(8, input.container.clientHeight / rows)
-  }
-
-  // Ghostty registers `canvas.addEventListener("mousedown", () => …focus())`
-  // unconditionally (no button check) — meant for desktop mouse clicks, but
-  // Chromium also synthesizes a compatibility `mousedown` from an unhandled
-  // touch sequence for legacy web compat. `preventDefault` on `touchstart` is
-  // the standard way to suppress that synthesis; kept as defense in depth,
-  // though it turned out NOT to be the actual cause of scroll reopening the
-  // keyboard (see below).
-  //
-  // Root cause, confirmed on-device: dismissing the keyboard (back button /
-  // tap outside) never actually calls `.blur()` on the hidden textarea — it
-  // only hides the IME UI. `document.activeElement` stays the textarea
-  // (verified: `vvHeight` back to full/no-keyboard while `activeElement` is
-  // still the terminal's textarea). Android's InputMethodManager can then
-  // re-show the keyboard on the NEXT touch of that still-focused view
-  // entirely at the platform level — independent of touchend, mousedown, or
-  // any `preventDefault()`, which is why blocking those JS events alone never
-  // stopped it. Blurring on every touchstart removes DOM focus before that
-  // native reshow can act; Ghostty's own touchend handler still calls
-  // `.focus()` for a genuine tap (mode never reaches "swipe"), so tap-to-open
-  // is unaffected — only scrolling now stays blurred throughout.
-  const suppressSyntheticMouseEvents = (e: TouchEvent) => {
-    e.preventDefault()
-    input.term.textarea?.blur()
-  }
-  const touchStartOptions: AddEventListenerOptions = { capture: true, passive: false }
-  input.container.addEventListener("touchstart", suppressSyntheticMouseEvents, touchStartOptions)
-  input.cleanups.push(() =>
-    input.container.removeEventListener("touchstart", suppressSyntheticMouseEvents, touchStartOptions),
-  )
-
-  const onTouchDownCapture = (e: PointerEvent) => {
-    if (e.pointerType !== "touch" || currentTouch) return
-    currentTouch = { id: e.pointerId, x: e.clientX, y: e.clientY, mode: "pending", scrollApplied: 0 }
-    armStuckTouchWatchdog()
-  }
-
-  const onTouchMoveCapture = (e: PointerEvent) => {
-    if (!currentTouch || e.pointerId !== currentTouch.id) return
-    armStuckTouchWatchdog()
-    const dy = e.clientY - currentTouch.y
-
-    if (currentTouch.mode === "pending") {
-      if (Math.hypot(e.clientX - currentTouch.x, dy) < MOBILE_SWIPE_THRESHOLD_PX) return
-      currentTouch.mode = "swipe"
-    }
-
-    // Swipe mode: consume the event so the surrounding app scroller does
-    // not also pan. ghostty clamps scrollLines at the buffer edges.
-    e.preventDefault()
-    e.stopPropagation()
-    // Drag-down = walking back in history = scroll UP (negative delta).
-    const targetRowsFromStart = Math.round(-dy / mobileCharHeight())
-    const totalDelta = targetRowsFromStart - currentTouch.scrollApplied
-    currentTouch.scrollApplied = targetRowsFromStart
-    if (totalDelta === 0) return
-
-    const charHeight = mobileCharHeight()
-    let delta = totalDelta
-    if (delta < 0 && overscrollPx > 0) {
-      // Dragging back toward history: unwind the fake overscroll first so
-      // the gesture feels continuous instead of jumping straight into real
-      // scrollback while the canvas is still shifted up.
-      const rowsToUnwind = Math.min(-delta, overscrollPx / charHeight)
-      overscrollPx = Math.max(0, overscrollPx - rowsToUnwind * charHeight)
-      delta += rowsToUnwind
-      applyOverscroll()
-    }
-    if (delta === 0) return
-
-    const beforeY = input.term.getViewportY()
-    input.term.scrollLines(delta)
-    if (delta > 0 && input.term.getViewportY() === 0) {
-      // Requested more "toward the bottom" scroll than the real buffer had
-      // left (already at viewportY 0) — the leftover becomes overscroll.
-      const unusedRows = delta - (beforeY - input.term.getViewportY())
-      if (unusedRows > 0) {
-        const maxOverscrollPx = input.container.clientHeight * MAX_OVERSCROLL_FACTOR
-        overscrollPx = Math.min(maxOverscrollPx, overscrollPx + unusedRows * charHeight)
-        applyOverscroll()
-      }
-    }
-  }
-
-  const onTouchEndOrCancel = (e: PointerEvent) => {
-    if (!currentTouch || e.pointerId !== currentTouch.id) return
-    disarmStuckTouchWatchdog()
-    lastGestureWasSwipe = currentTouch.mode === "swipe"
-    currentTouch = null
-    // DO NOT preventDefault/stopPropagation here — Ghostty's native
-    // touchend handler on the canvas must still fire so the IME attaches
-    // correctly to the textarea. This is the lesson from the 2026-04-23
-    // regression where blocking touchend left the softkeyboard visually
-    // open but keystrokes never reached the textarea.
-  }
-
-  const touchCaptureOptions: AddEventListenerOptions = { capture: true }
-  const touchMoveOptions: AddEventListenerOptions = { capture: true, passive: false }
-  input.container.addEventListener("pointerdown", onTouchDownCapture, touchCaptureOptions)
-  input.container.addEventListener("pointermove", onTouchMoveCapture, touchMoveOptions)
-  input.container.addEventListener("pointerup", onTouchEndOrCancel, touchCaptureOptions)
-  input.container.addEventListener("pointercancel", onTouchEndOrCancel, touchCaptureOptions)
-  input.cleanups.push(() => {
-    input.container.removeEventListener("pointerdown", onTouchDownCapture, touchCaptureOptions)
-    input.container.removeEventListener("pointermove", onTouchMoveCapture, touchMoveOptions)
-    input.container.removeEventListener("pointerup", onTouchEndOrCancel, touchCaptureOptions)
-    input.container.removeEventListener("pointercancel", onTouchEndOrCancel, touchCaptureOptions)
-  })
-
-  // Prevent Ghostty's `canvas.addEventListener("touchend", g.focus())` ONLY
-  // when the gesture was a swipe — so scrolling never toggles the
-  // softkeyboard state. For taps (mode stays "pending"), we let touchend
-  // bubble to the canvas so Ghostty attaches the Android IME normally.
-  // This conditional block is safe where the v3.2 attempt (unconditional
-  // stopImmediatePropagation on touchend) was not. Reads `lastGestureWasSwipe`
-  // (captured synchronously in onTouchEndOrCancel's pointerup/pointercancel
-  // handler) rather than `currentTouch?.mode`, which is always already null
-  // by the time this touchend handler runs.
-  //
-  // Also blocks while `currentTouch` is still non-null (another pointer is
-  // still actively tracked): confirmed on-device that Android reports a
-  // second, ~4ms-lived pointer (its own distinct pointerId) partway through
-  // a real one-finger scroll. `onTouchDownCapture` already ignores that
-  // second pointerdown (`if (... || currentTouch) return`), so its matching
-  // pointerup never reaches `onTouchEndOrCancel` either (pointerId mismatch)
-  // and `lastGestureWasSwipe` is never updated for it — leaving this touchend
-  // to fall through on a stale value from whatever gesture came before.
-  // Any touchend arriving while a swipe is still in progress is spurious by
-  // definition (a real tap-to-focus never overlaps another active touch).
-  const blockTouchEndIfSwipe = (e: TouchEvent) => {
-    if (lastGestureWasSwipe || currentTouch) {
-      e.stopPropagation()
-    }
-  }
-  const touchBlockerOptions: AddEventListenerOptions = { capture: true, passive: true }
-  input.container.addEventListener("touchend", blockTouchEndIfSwipe, touchBlockerOptions)
-  input.container.addEventListener("touchcancel", blockTouchEndIfSwipe, touchBlockerOptions)
-  input.cleanups.push(() => {
-    input.container.removeEventListener("touchend", blockTouchEndIfSwipe, touchBlockerOptions)
-    input.container.removeEventListener("touchcancel", blockTouchEndIfSwipe, touchBlockerOptions)
-  })
-
-  input.container.addEventListener("click", input.handleLinkClick, {
-    capture: true,
-  })
-  input.cleanups.push(() =>
-    input.container.removeEventListener("click", input.handleLinkClick, {
-      capture: true,
-    }),
-  )
-
-  input.term.textarea?.addEventListener("focus", handleTextareaFocus)
-  input.term.textarea?.addEventListener("blur", handleTextareaBlur)
-  input.cleanups.push(() => input.term.textarea?.removeEventListener("focus", handleTextareaFocus))
-  input.cleanups.push(() => input.term.textarea?.removeEventListener("blur", handleTextareaBlur))
-}
-
 const persistTerminal = (input: {
   term: Term | undefined
   addon: SerializeAddon | undefined
@@ -567,7 +279,16 @@ export const Terminal = (props: TerminalProps) => {
     console.info("[terminal-debug]", msg)
   }
   let container!: HTMLDivElement
-  const [local, others] = splitProps(props, ["pty", "class", "classList", "autoFocus", "onConnect", "onConnectError", "onSend"])
+  const [local, others] = splitProps(props, [
+    "pty",
+    "class",
+    "classList",
+    "autoFocus",
+    "onConnect",
+    "onConnectError",
+    "onSend",
+    "onSelectionApi",
+  ])
   const id = local.pty.id
   const probe = terminalProbe(id)
   const restore = typeof local.pty.buffer === "string" ? local.pty.buffer : ""
@@ -1259,6 +980,17 @@ export const Terminal = (props: TerminalProps) => {
       }
       local.onSend?.(sendBytes)
       cleanups.push(() => local.onSend?.(undefined))
+
+      local.onSelectionApi?.({
+        hasSelection: () => t.getSelection().length > 0,
+        copySelection: () => t.copySelection(),
+        paste: (text) => t.paste(text),
+        onSelectionChange: (cb) => {
+          const disposable = t.onSelectionChange(cb)
+          return () => disposeIfDisposable(disposable)
+        },
+      })
+      cleanups.push(() => local.onSelectionApi?.(undefined))
 
       // Android suspends WebView JS timers/networking while the app is
       // backgrounded (screen lock, app switch). If the PTY WebSocket dies
